requirements.txt
openai>=1.0.0
markdown>=3.4.0
pyyaml>=6.0
tqdm>=4.0.0
requests>=2.32.3
pre-commit>=3.5.0
black>=23.11.0
flake8>=6.1.0
together>=0.2.11
anthropic>=0.49.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
google-genai>=0.8.5
google-generativeai>=0.8.5
google-cloud-storage>=3.1.0
statsmodels>=0.14.4
krippendorff>=0.8.1
flask
pyyaml
scipy
